requirements.txt
anthropic==0.40.0
python-dotenv==1.0.1
google-cloud-texttospeech==2.21.1
playsound==1.3.0
